pyproject.toml
[project]
name = "repl-nix-visionaryexecutor"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "pillow>=11.0.0",
    "requests>=2.32.3",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.0.1",
    "werkzeug",
]
